requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Admin==1.6.1
Werkzeug>=3.0.0
click==8.1.3
gunicorn==20.1.0
gevent==22.10.2
pytest==7.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.4.2
reportlab==4.0.9

